+++ hypervisor/include/common/ptdev.h
@@ -8,7 +8,6 @@
 #define PTDEV_H
 #include <list.h>
 #include <spinlock.h>
-#include <pci.h>
 #include <timer.h>
 
 #define PTDEV_INTR_MSI		(1U << 0U)
@@ -24,16 +23,6 @@ union source_id (name) = {.msi_id = {.bdf = (a), .entry_nr = (b)} }
 
 #define DEFINE_IOAPIC_SID(name, a, b)	\
 union source_id (name) = {.intx_id = {.pin = (a), .src = (b)} }
-
-union source {
-	uint16_t ioapic_id;
-	union pci_bdf msi;
-};
-
-struct intr_source {
-	bool is_msi;
-	union source src;
-};
 
 union irte_index {
 	uint16_t index;
